refactor(edge): extract quic listener startup and bind orchestration
